requirements.txt
pandas==1.3.5
numpy
tqdm
matplotlib
scikit-learn
goatools
seaborn==0.11.2
memory-profiler
ipykernel
python-dotenv
neo4j